requirements.txt
Django==2.2.10
djangorestframework==3.9.1
gunicorn==19.9.0
psycopg2==2.7.7
psycopg2-binary==2.7.7
whitenoise
Pillow